fix: e2e tests (#3487)
## Overview

Fixes #3453 

I encountered a lot of flakiness while running knuu tests in ci(view
logs in issue description) to resolve them i added a few workarounds
with @smuu which broke other things in our existing suite. those changes
are now reverted since the tests started passing on main again after
scaling some of the clusters that were stuck. DevOps is working on
addressing flakiness on knuu.

E2E tests are green:
https://github.com/celestiaorg/celestia-app/actions/runs/9179315870/job/25241185362?pr=3487

My changes:
-  Fix git tag fetching issue
-  Extract e2e tests into its own yaml file
-  Separate MinorVersionCompatibility and MajorupgradeToV2
