+++ .github/workflows/nightly.yml
@@ -7,22 +7,8 @@ on:
 
 jobs:
   test-e2e:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-go@v5
-        with:
-          go-version-file: 'go.mod'
-
-      - name: Setup kubeconfig
-        env:
-          KUBECONFIG_FILE: ${{ secrets.KNUU_KUBECONFIG_FILE }}
-        run: |
-          mkdir -p $HOME/.kube
-          echo "${KUBECONFIG_FILE}" > $HOME/.kube/config
-      - name: Run e2e tests
-        run: make test-e2e
+    uses: ./.github/workflows/test-e2e.yml
+    secrets: inherit
 
   test:
     uses: ./.github/workflows/test.yml
